requirements.txt
aliyun-python-sdk-core-v3==2.13.11
asgiref==3.2.3
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==3.0.1
django-cors-headers==3.2.0
django-filter==2.2.0
django-guardian==2.1.0
django-mdeditor==0.1.18
django-redis==4.11.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
drf-yasg==1.17.0
idna==2.8
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.3
jmespath==0.9.4
MarkupSafe==1.1.1
mixins==0.1.1
mysqlclient==1.4.6
packaging==19.2
pycryptodome==3.9.4
PyJWT==1.7.1
pyparsing==2.4.6
pytz==2019.3
redis==3.3.11
requests==2.22.0
ruamel.yaml==0.16.5
ruamel.yaml.clib==0.2.0
six==1.13.0
sqlparse==0.3.0
uritemplate==3.0.1
urllib3==1.25.7
